Fixed CthInitialize bug on uth machines.

+++ src/ck-core/env_macros.h
@@ -12,7 +12,10 @@
  * REVISION HISTORY:
  *
  * $Log$
- * Revision 2.13  1997-03-24 23:10:13  milind
+ * Revision 2.14  1997-10-29 23:52:46  milind
+ * Fixed CthInitialize bug on uth machines.
+ *
+ * Revision 2.13  1997/03/24 23:10:13  milind
  * Made envelope 64-bit safe by replacing an int with a void *.
  *
  * Revision 2.12  1997/02/06 19:52:44  jyelon
@@ -133,7 +136,7 @@ typedef struct envelope {
 /*********************************************************/
 /** Arrangement for i_tag2                              **/
 /*********************************************************/
-#define GetEnv_count(e)		        ((int)(((ENVELOPE *)(e))->i_tag2))
+#define GetEnv_count(e)		        ((size_t)(((ENVELOPE *)(e))->i_tag2))
 #define SetEnv_count(e,x)		(((ENVELOPE *)(e))->i_tag2=((void *)(x)))
 
 #define GetEnv_chareBlockPtr(e)	        ((CHARE_BLOCK *)(((ENVELOPE *)(e))->i_tag2))
@@ -142,7 +145,7 @@ typedef struct envelope {
 #define SetEnv_vidBlockPtr(e,x)	        (((ENVELOPE *)(e))->i_tag2=(x))
 #define GetEnv_vidBlockPtr(e)		((CHARE_BLOCK *)(((ENVELOPE *)(e))->i_tag2))
 
-#define GetEnv_boc_num(e) 		((int)(((ENVELOPE *)(e))->i_tag2))
+#define GetEnv_boc_num(e) 		((size_t)(((ENVELOPE *)(e))->i_tag2))
 #define SetEnv_boc_num(e,x) 		(((ENVELOPE *)(e))->i_tag2=((void *)(x)))
 
 /*********************************************************/
